Update and rename test_cases.py to test_calc.py
