+++ tests/test_calc.py
@@ -1,0 +1,22 @@
+# test_cases.py
+
+import pytest
+
+# Список тестовых примеров и ожидаемых результатов
+test_cases = [
+    ("12 + 15 / (3 + 2) * 4", 12 + 15 / (3 + 2) * 4),
+    ("(2 + 3) * 5", (2 + 3) * 5),
+    ("(8 / 4) + (3 * 2)", (8 / 4) + (3 * 2)),
+    ("(10 + 2) / 3", (10 + 2) / 3),
+    ("3 * (4 + 2) - 5", 3 * (4 + 2) - 5),
+    ("test", 'Ошибка ввода. Введите выражение аналогично примеру'),
+    ("(32+3", 'Ошибка ввода. Закройте все скобки.'),
+    ("12 / 0", '*** Ошибка: деление на ноль ***'),
+    ("", 'Ошибка ввода. Введите выражение аналогично примеру'),
+    ("((((31*12)-15/3)+2^3)-5)", ((((31*12)-15/3)+2**3)-5)),
+]
+
+@pytest.mark.parametrize("expression, expected", test_cases)
+def test_calc(expression, expected):
+    from service.service_functions import calc
+    assert calc(expression) == expected
